Move test helpers to a submodule with feature.
