--- ethcore/hbbft_engine/tests/p2p.rs
@@ -1,20 +1,0 @@
-extern crate ethcore;
-extern crate hbbft_engine;
-extern crate inventory;
-
-use ethcore::spec::Spec;
-
-#[test]
-fn test_nodes_p2p() {
-	hbbft_engine::init();
-
-	// Load the chain specification.
-	let spec = Spec::load(
-		&::std::env::temp_dir(),
-		include_bytes!("../res/honey_badger_bft.json") as &[u8],
-	)
-	.expect(concat!("Chain spec is invalid."));
-
-	// The engine was created from the spec.
-	let _engine = spec.engine;
-}
